pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yupoo-scraper"
version = "0.1.0"
description = "Yupoo 产品采集 → 图片处理 → 查重 → 微店上架"
requires-python = ">=3.10"
dependencies = [
    # GUI
    "PySide6>=6.6.0,<7.0.0",
    "qasync>=0.27.0,<1.0.0",
    # 采集
    "aiohttp>=3.9.0,<4.0.0",
    "yarl>=1.9.0,<2.0.0",
    "beautifulsoup4>=4.12.0,<5.0.0",
    "lxml>=5.0.0,<6.0.0",
    # 图片处理
    "Pillow>=10.0.0,<12.0.0",
    "imagehash>=4.3.0,<5.0.0",
    "rembg>=2.0.50,<2.0.58",
    # 上架
    "playwright>=1.40.0,<2.0.0",
    # ML (产品拆分 + 查重)
    "numpy>=1.24.0,<3.0.0",
    "torch>=2.0.0,<3.0.0",
    "torchvision>=0.15.0,<1.0.0",
    "open-clip-torch>=2.24.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "opencv-python-headless>=4.8.0,<5.0.0",
    "faiss-cpu>=1.7.4,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
yupoo-scraper = "yupoo_scraper.main:main"

[tool.setuptools.packages.find]
include = ["yupoo_scraper*"]
